requirements.txt
streamlit
python-dotenv
PyMuPDF
requests